Simplified code. Reduced duplications.
