+++ pyEDAA/OutputFilter/Xilinx/PlaceDesign.py
@@ -32,315 +32,9 @@
 from typing import Generator, ClassVar, List, Type, Dict, Tuple
 
 from pyTooling.Decorators  import export
-from pyTooling.MetaClasses import ExtendedType
 
 from pyEDAA.OutputFilter.Xilinx           import Line, VivadoMessage, LineKind
-from pyEDAA.OutputFilter.Xilinx.Exception import ProcessorException
-from pyEDAA.OutputFilter.Xilinx.Common2   import BaseParser, VivadoMessagesMixin
-
-
-@export
-class Task(BaseParser, VivadoMessagesMixin, metaclass=ExtendedType, slots=True):
-	# _START:  ClassVar[str]
-	# _FINISH: ClassVar[str]
-	_TIME:   ClassVar[str] = "Time (s):"
-
-	_command:  "Command"
-	_duration: float
-
-	def __init__(self, command: "Command"):
-		super().__init__()
-		VivadoMessagesMixin.__init__(self)
-
-		self._command = command
-
-	def _TaskStart(self, line: Line) -> Generator[Line, Line, Line]:
-		if not line.StartsWith(self._START):
-			raise ProcessorException()
-
-		line._kind = LineKind.TaskStart
-		nextLine = yield line
-		return nextLine
-
-	def _TaskFinish(self, line: Line) -> Generator[Line, Line, Line]:
-		if not line.StartsWith(self._FINISH):
-			raise ProcessorException()
-
-		line._kind = LineKind.TaskEnd
-		line = yield line
-		while self._TIME is not None:
-			if line.StartsWith(self._TIME):
-				line._kind = LineKind.TaskTime
-				break
-
-			line = yield line
-
-		line = yield line
-		return line
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		line = yield from self._TaskStart(line)
-
-		while True:
-			if line._kind is LineKind.Empty:
-				line = yield line
-				continue
-			elif line.StartsWith("Ending"):
-				break
-			elif isinstance(line, VivadoMessage):
-				self._AddMessage(line)
-			elif line.StartsWith(self._TIME):
-				line._kind = LineKind.TaskTime
-				nextLine = yield line
-				return nextLine
-
-			line = yield line
-
-		nextLine = yield from self._TaskFinish(line)
-		return nextLine
-
-
-@export
-class Phase(BaseParser, VivadoMessagesMixin, metaclass=ExtendedType, slots=True):
-	# _START:  ClassVar[str]
-	# _FINISH: ClassVar[str]
-
-	_task:     Task
-	_duration: float
-
-	def __init__(self, task: Task):
-		super().__init__()
-		VivadoMessagesMixin.__init__(self)
-
-		self._task = task
-
-	def _PhaseStart(self, line: Line) -> Generator[Line, Line, Line]:
-		if not line.StartsWith(self._START):
-			raise ProcessorException()
-
-		line._kind = LineKind.PhaseStart
-		nextLine = yield line
-		return nextLine
-
-	def _PhaseFinish(self, line: Line) -> Generator[Line, Line, None]:
-		if not line.StartsWith(self._FINISH):
-			raise ProcessorException()
-
-		line._kind = LineKind.PhaseEnd
-		line = yield line
-
-		if self._TIME is not None:
-			while self._TIME is not None:
-				if line.StartsWith(self._TIME):
-					line._kind = LineKind.PhaseTime
-					break
-
-				line = yield line
-
-			line = yield line
-
-		if self._FINAL is not None:
-			while self._FINAL is not None:
-				if line.StartsWith(self._FINAL):
-					line._kind = LineKind.PhaseFinal
-					break
-
-				line = yield line
-
-			line = yield line
-
-		return line
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		line = yield from self._PhaseStart(line)
-
-		while True:
-			if line._kind is LineKind.Empty:
-				line = yield line
-				continue
-			elif line.StartsWith(self._FINISH):
-				break
-			elif isinstance(line, VivadoMessage):
-				self._AddMessage(line)
-
-			line = yield line
-
-		nextLine = yield from self._PhaseFinish(line)
-		return nextLine
-
-
-@export
-class SubPhase(BaseParser, VivadoMessagesMixin, metaclass=ExtendedType, slots=True):
-	# _START:  ClassVar[str]
-	# _FINISH: ClassVar[str]
-
-	_phase:    Phase
-	_duration: float
-
-	def __init__(self, phase: Phase):
-		super().__init__()
-		VivadoMessagesMixin.__init__(self)
-
-		self._phase = phase
-
-	def _SubPhaseStart(self, line: Line) -> Generator[Line, Line, Line]:
-		if not line.StartsWith(self._START):
-			raise ProcessorException()
-
-		line._kind = LineKind.SubPhaseStart
-		nextLine = yield line
-		return nextLine
-
-	def _SubPhaseFinish(self, line: Line) -> Generator[Line, Line, None]:
-		if not line.StartsWith(self._FINISH):
-			raise ProcessorException()
-
-		if self._TIME is None:
-			line._kind = LineKind.SubPhaseTime
-		else:
-			line._kind = LineKind.SubPhaseEnd
-
-			line = yield line
-			while self._TIME is not None:
-				if line.StartsWith(self._TIME):
-					line._kind = LineKind.SubPhaseTime
-					break
-
-				line = yield line
-
-		nextLine = yield line
-		return nextLine
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		line = yield from self._SubPhaseStart(line)
-
-		while True:
-			if line._kind is LineKind.Empty:
-				line = yield line
-				continue
-			elif line.StartsWith(self._FINISH):
-				break
-			elif isinstance(line, VivadoMessage):
-				self._AddMessage(line)
-
-			line = yield line
-
-		nextLine = yield from self._SubPhaseFinish(line)
-		return nextLine
-
-
-@export
-class SubSubPhase(BaseParser, VivadoMessagesMixin, metaclass=ExtendedType, slots=True):
-	# _START:  ClassVar[str]
-	# _FINISH: ClassVar[str]
-
-	_subphase: SubPhase
-	_duration: float
-
-	def __init__(self, subphase: SubPhase):
-		super().__init__()
-		VivadoMessagesMixin.__init__(self)
-
-		self._subphase = subphase
-
-	def _SubSubPhaseStart(self, line: Line) -> Generator[Line, Line, Line]:
-		if not line.StartsWith(self._START):
-			raise ProcessorException()
-
-		line._kind = LineKind.SubSubPhaseStart
-		nextLine = yield line
-		return nextLine
-
-	def _SubSubPhaseFinish(self, line: Line) -> Generator[Line, Line, None]:
-		if not line.StartsWith(self._FINISH):
-			raise ProcessorException()
-
-		line._kind = LineKind.SubSubPhaseEnd
-		line = yield line
-
-		while self._TIME is not None:
-			if line.StartsWith(self._TIME):
-				line._kind = LineKind.SubSubPhaseTime
-				break
-
-			line = yield line
-
-		nextLine = yield line
-		return nextLine
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		line = yield from self._SubSubPhaseStart(line)
-
-		while True:
-			if line._kind is LineKind.Empty:
-				line = yield line
-				continue
-			elif line.StartsWith(self._FINISH):
-				break
-			elif isinstance(line, VivadoMessage):
-				self._AddMessage(line)
-
-			line = yield line
-
-		nextLine = yield from self._SubSubPhaseFinish(line)
-		return nextLine
-
-
-@export
-class SubSubSubPhase(BaseParser, VivadoMessagesMixin, metaclass=ExtendedType, slots=True):
-	# _START:  ClassVar[str]
-	# _FINISH: ClassVar[str]
-
-	_subsubphase: SubSubPhase
-	_duration: float
-
-	def __init__(self, subsubphase: SubSubPhase):
-		super().__init__()
-		VivadoMessagesMixin.__init__(self)
-
-		self._subsubphase = subsubphase
-
-	def _SubSubSubPhaseStart(self, line: Line) -> Generator[Line, Line, Line]:
-		if not line.StartsWith(self._START):
-			raise ProcessorException()
-
-		line._kind = LineKind.SubSubSubPhaseStart
-		nextLine = yield line
-		return nextLine
-
-	def _SubSubSubPhaseFinish(self, line: Line) -> Generator[Line, Line, None]:
-		if not line.StartsWith(self._FINISH):
-			raise ProcessorException()
-
-		line._kind = LineKind.SubSubSubPhaseEnd
-		line = yield line
-
-		while self._TIME is not None:
-			if line.StartsWith(self._TIME):
-				line._kind = LineKind.SubSubSubPhaseTime
-				break
-
-			line = yield line
-
-		nextLine = yield line
-		return nextLine
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		line = yield from self._SubSubSubPhaseStart(line)
-
-		while True:
-			if line._kind is LineKind.Empty:
-				line = yield line
-				continue
-			elif line.StartsWith(self._FINISH):
-				break
-			elif isinstance(line, VivadoMessage):
-				self._AddMessage(line)
-
-			line = yield line
-
-		nextLine = yield from self._SubSubSubPhaseFinish(line)
-		return nextLine
+from pyEDAA.OutputFilter.Xilinx.Common2   import Task, Phase, SubPhase, SubSubPhase, SubSubSubPhase
 
 
 @export
@@ -1016,13 +710,6 @@ class PlacerTask(Task):
 		Phase3_DetailPlacement,
 		Phase4_PostPlacementOptimizationAndCleanUp
 	)
-
-	_phases: Dict[Type[Phase], Phase]
-
-	def __init__(self, command: "Command"):
-		super().__init__(command)
-
-		self._phases = {p: p(self) for p in self._PARSERS}
 
 	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
 		line = yield from self._TaskStart(line)
